+++ src/Examples.EntityFrameworkCore.SqlServer.Tests/ContosoUniversity.SqlServer.Tests/ContosoUniversityFixture.cs
@@ -10,6 +10,7 @@ namespace Examples.ContosoUniversity.SqlServer.Tests;
 
 public class ContosoUniversityFixture : IDisposable
 {
+    public static bool Enabled => !string.IsNullOrEmpty(Environment.GetEnvironmentVariable("MSSQL_SERVICE"));
     private static readonly Lock _lock = new();
     private static bool _databaseInitialized;
 
@@ -26,12 +27,7 @@ public class ContosoUniversityFixture : IDisposable
                 .AddFilter("ContosoUniversity", LogLevel.Trace)
                 .AddDelegateLogger(x => _logging?.Invoke($"[{DateTime.Now:HH:mm:ss.fff}] {x.LogLevel}: {x.Message}")));
 
-        var connectionString = GetConnectionString();
-        services.AddDbContext<SchoolContext>(builder
-            => builder.UseSqlServer(connectionString)
-        );
-
-        services.AddScoped<IStudentRepository, StudentRepository>();
+        ConfigureServices(services);
 
         _serviceProvider = services.BuildServiceProvider();
 
@@ -72,8 +68,28 @@ public class ContosoUniversityFixture : IDisposable
         return connectionString;
     }
 
+    private static void ConfigureServices(ServiceCollection services)
+    {
+        if (!Enabled)
+        {
+            return;
+        }
+
+        var connectionString = GetConnectionString();
+        services.AddDbContext<SchoolContext>(builder
+            => builder.UseSqlServer(connectionString)
+        );
+
+        services.AddScoped<IStudentRepository, StudentRepository>();
+    }
+
     private void InitializeDatabase()
     {
+        if (!Enabled)
+        {
+            return;
+        }
+
         var context = ServiceProvider.GetRequiredService<SchoolContext>();
 
         context.Database.EnsureDeleted();
